:recycle: refactor transformable_box.dart file
