+++ packages/flutter_box_transform/lib/flutter_box_transform.dart
@@ -2,9 +2,10 @@ library flutter_box_transform;
 
 export 'package:box_transform/box_transform.dart';
 
-export 'src/extensions.dart' hide DimensionExt;
-export 'src/handle.dart';
+export 'src/extensions.dart';
+export 'src/handles.dart';
 export 'src/transformable_box.dart';
 export 'src/transformable_box_controller.dart';
+export 'src/typedefs.dart';
 export 'src/ui_box_transform.dart';
 export 'src/ui_result.dart';
